Fixed a potential bug
There was a possibility that because we handled errors by setting the
pid target to current position, the error would become zero- potentially
giving the go-ahead on the calibration state. We now disable the loop on
the no valid target error, and check if the loop is enabled during the
calibration state.